[NFC] Satisfy frontend clang-tidy on the arena code
Explicit casts for the slab-size constant, placement-new void*, and
Shape narrowing; add missing <cstdint> in TypedASTPrinter

+++ lib/AST/TypedASTPrinter.cpp
@@ -1,4 +1,5 @@
 #include "trsc/AST/TypedASTPrinter.h"
+#include <cstdint>
 #include <iomanip>
 
 namespace trsc {
@@ -518,8 +519,8 @@ void TypedASTPrinter::visitArrayExpr(ArrayExpr *Node) {
   }
   OS << "]";
 
-  int Total = 1;
-  for (int D : Shape)
+  int64_t Total = 1;
+  for (int64_t D : Shape)
     Total *= D;
   OS << " (" << Total << " elements)";
   OS << "\n";
